Lazily load the library + add suspense for when the library is loading
